src/main/python/scoring/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pt-matsim-scoring"
version = "1.0.0"
description = "Level of Service (LOS) scoring module for public transit analysis based on MATSim data"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Minh Hai Hoang"}
]
keywords = ["matsim", "public-transit", "los", "level-of-service", "transit-scoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
    "pandas>=2.0.0",
    "PyYAML>=6.0",
    "lxml>=4.0.0",
    "pyarrow>=14.0.0",
]

[project.optional-dependencies]
dev = [
    "build",
    "twine",
    "pytest",
]

[tool.setuptools]
packages = ["scoring"]
package-dir = {"scoring" = "."}
